Schedule description is now displayed on flow page
